Adding feedback to the rake command. [Solving issue #1]

+++ spec/lib/auto_tag_version/utils_spec.rb
@@ -6,18 +6,21 @@ describe AutoTagVersion do
   end
 
   module Rails; def self.application; end; end
-  module APP; end;
+  module APP; end
 
   let(:tag_version) { "#{rand(9)}.#{rand(9)}.#{rand(9)}" }
   let(:app) { APP }
   let(:app_version_file) { "config/initializers/app_version.rb" }
+  let(:git_add_cmd) { "git add #{app_version_file} && git commit -m \"Bumping version #{tag_version}\" && git tag #{tag_version}" }
+
 
   before do
     allow(Rails).to receive_message_chain("application.class.parent_name").and_return(app.to_s)
+    allow(subject).to receive(:`).with(git_add_cmd)
   end
 
   after do
-    FileUtils.rm(app_version_file)
+    FileUtils.rm(app_version_file) if File.exist?(app_version_file)
   end
 
   describe ".tag!" do
@@ -33,12 +36,28 @@ describe AutoTagVersion do
     end
 
     context "git integration" do
-      let(:git_msg) { "git add #{app_version_file} && git commit -m \"Bumping version #{tag_version}\" && git tag #{tag_version}" }
-
       it "should run the correct git command" do
-        expect(subject).to receive(:`).with(git_msg)
+        expect(subject).to receive(:`).with(git_add_cmd)
         subject.tag!(tag_version)
       end
+    end
+  end
+
+  describe ".print_last_tag_information" do
+    let(:git_tags_cmd) { "git describe --abbrev=0 --tags" }
+
+    before do
+      allow(subject).to receive(:`).with(git_tags_cmd).and_return(tag_version)
+      subject.tag!(tag_version)
+    end
+
+    it "should print successful message" do
+      expect { subject.print_last_tag_information }.to output("Everything OK! Last git tag created: #{tag_version}\n").to_stdout
+    end
+
+    it "should print successful message" do
+      allow(subject).to receive(:`).with(git_tags_cmd).and_return("test.0.1")
+      expect { subject.print_last_tag_information }.to output("Something goes wrong with the tag creation. Check your git log.\n").to_stdout
     end
   end
 end
